requirements.txt
asgiref==3.5.0
Django==4.0.3
django-environ==0.8.1
django-extensions==3.1.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
docopt==0.6.2
num2words==0.5.10
PyJWT==2.3.0
pytz==2022.1
sqlparse==0.4.2
